moved tests out of __tests__
